Index editions updated_at
EditionFilter is used to find and paginate editions on the editions
index page. It sorts editions descending on updated_at, which has
been the behaviour since 2011 (24b2a027). There's never been an index on this column.

For the database to find the most recent editions for any query, it
always has to do a full table scan.

If we compare two `select * from editions where ... order by COLUMN limit 25`
queries, the first one using the `updated_at` column which has no index,
and the second one using the `public_timestamp` column which has we can
see the performance difference on the table.

In integration:

    EXPLAIN ANALYZE SELECT * FROM `editions` WHERE `editions`.`state` != 'deleted' ORDER BY editions.updated_at DESC LIMIT 25 OFFSET 0
    -> Limit: 25 row(s)  (cost=134443 rows=25) (actual time=4847..4847 rows=25 loops=1)
        -> Sort: editions.updated_at DESC, limit input to 25 row(s) per chunk  (cost=134443 rows=1.31e+6) (actual time=4847..4847 rows=25 loops=1)
            -> Filter: (editions.state <> 'deleted')  (cost=134443 rows=1.31e+6) (actual time=0.767..4035 rows=1.28e+6 loops=1)
                -> Table scan on editions  (cost=134443 rows=1.31e+6) (actual time=0.0331..3810 rows=1.43e+6 loops=1)

    EXPLAIN ANALYZE SELECT * FROM `editions` WHERE `editions`.`state` != 'deleted' ORDER BY editions.public_timestamp DESC LIMIT 25 OFFSET 0
    -> Limit: 25 row(s)  (cost=2.59 rows=24.5) (actual time=0.0439..0.17 rows=25 loops=1)
        -> Filter: (editions.state <> 'deleted')  (cost=2.59 rows=24.5) (actual time=0.0432..0.167 rows=25 loops=1)
            -> Index scan on editions using index_editions_on_public_timestamp (reverse)  (cost=2.59 rows=49) (actual time=0.0413..0.161 rows=25 loops=1)

Note the difference in actual time between the Table scan (3810ms) and
the Index scan (0.161ms) - 23,000 times faster.

In practice though we don't want to sort by public_timestamp, we
actually do want to sort by updated_at. So the right thing to do seems
to be to add an index to this column.

I haven't yet tested:
- how long this index will take to add initially
- whether it will lock the table while it's being added
- what effect it will have on insert performance (I assume minimal)
- how much extra storage the index will take up (I assume minimal again - should be no more than the other similar indexes on this table)
- what the real world effect will be on page loads - this isn't the only
  slow query on the editions index page

For all of these things, I think the best option is to run the migration
in integration and see what happens.