Fix back links for Organisations
"Pages" (or "Corporate information pages") and "Social media
accounts" appear on both Organisation and WorldwideOrganisation
index pages. The back links for these made an assumption that they
only appear on worldwide organisation pages, meaning the back link
was incorrect for those.

I considered adding a test for this, but it doesn't seem worth
the test bloat. If it breaks, it's not the end of the world.

Zendesk: https://govuk.zendesk.com/agent/tickets/5655143