Add bulk fields to RepublishingEvents
This will be used to denote a "bulk" republishing task being run, and
passed to the republishing worker.

We're adding our first `bulk_content_type` here, but this list will grow
in the very near future so we've made it an `enum` that we'll extend in
future branches.

We found that existing `RepublishingEvent` records will be set to `bulk:
false`, which is accurate.

Co-authored-by: Ynda <yndajas@gmail.com>