Migrate non-editionable worldwide organisations to be editionable
This migrates non-editionable worldwide organisations and all their
attributes to be editionable.

An integrity check of the migrated content (comparing Publishing API
presenter output) is carried out for all things that are migrated. If
any part of the integrity checks fail, the entire transaction will
rollback and the migration will fail.

Note 1: this migration is irreversible.

Note 2: there are some legacy worldwide organisations that were created
prior to some validations being put in place, therefore they have
invalid values. The migration is skipping validations where necessary
(using different strategies, as required) to ensure the values are
unchanged when migrating over. Users will be alerted to the invalid
content next time they edit the document.

Note 3: we cannot use a data migration as they do not have access to
presenters and helpers that are needed to perform the consistency check
before commiting the transaction.