Ensure only drafts associated with document first published after the date provided are marked as political
There is no reason to mark draft editions for documents published before the specified date as political. This task is to be used to apply political status to documents for an organisation that was formed from an existing organisation and should have been marked political at that time. The date is expected to be the date from which the new organisation came into official being. We have generalised the task in case there are other uses in the future.