Render bulk republishing pages links on main index
We'll be able to add new pages to this easily by adding them to the
helper.

The tests for `bulk_content_type_index_rows` are a little brute force
here, I'm not 100% keen on them, but this all felt a bit too much logic
to have in the views and I think a helper is the best place of this,
seeing as we don't have a Presenter class.

I’d have liked to use Minitest::Spec::DSL here to structure the helper
tests a bit more nicely but it appears to be causing havoc with
including modules for some reason, so I’ve left it without for now. I
can't see the number of tests in this file growing all that much
larger, so maybe this is acceptable.

Co-authored-by: Ynda <yndajas@gmail.com>