Delete blank contacts
There is currently no way of deleting the final contact from a Worldwide
Organisation's office in the Whitehall user interface.

Therefore leaving the final one should delete the entry. There was a
test label for this, but the test didn't actual test that scenario.
Therefore adding a test for this and updating the controller to work in
that way.