Duplicate contact numbers on new edition
We are currently copying the office and the associated contact when
creating a new edition, however not the contact numbers.

This adds that duplication, so contact numbers are not lost when a new
edition is created.

We also need to save the record, as otherwise it will just be built but
not persisted in some cases.