Add default value for `RepublishingEvent`s
When we added the `bulk` field in
[`578590b27`](https://github.com/alphagov/whitehall/pull/9075/commits/578590b2781ce85e1bbcdac1822a0a11b1c0ee22)
we didn't set a default value as there were no `RepublishingEvent`s.

This caused Rubocop to fail on `main` due to us not providing a default
value for a new field for an existing database entity - see
[Rails/NotNullColumn
documentation](https://docs.rubocop.org/rubocop-rails/cops_rails.html#railsnotnullcolumn),
for more information.

For some reason this didn't warn us before we merged our PR to main. It
looks like `rubocop-govuk` was updated to 4.17.1 and `rubocop-rails` to
`2.25.0`, but this is an old change from 2017 so we're a little confused
as to why it's only now complaining.

Here we add an explicit `false` default value for `RepublishingEvent`s
and ignore the Rubocop violation on the previous migration.

I've had to use `up` rather than `change` here as `change_column` is not
reversible. According to a comment on the accepted answer on [this Stack
Overflow
post](https://github.com/alphagov/whitehall/pull/9075/commits/578590b2781ce85e1bbcdac1822a0a11b1c0ee22),

```
If you need reversible migrations, put this in an up block rather
than a change block. You can leave the down block empty. It won't revert
the table to the original condition but the migration can be rolled
back.
```

This is an intermediate step before setting this back to have no `null`
value to prevent accidentally setting `RepublishingEvent`s as non-`bulk`
when they're not in future changes to the code.