Present `auth_bypass_ids` for editionable worldwide organisation offices
These values were missed from the presenter. By sending them to
Publishing API, we are able to offer a shareable preview feature to
publishers.

We are using the `auth_bypass_id` from the parents worldwide
organisation to create a seemless journey where users only need a single
token to see all offices that are part of the organisation.

This presenter is used for worldwide offices that are associated with
both editionable and non-editionable worldwide organisations. Therefore
adding a test case for editionable worldwide organisations that was
missing.