Add a `details` column to Editions
At the moment, Editions are a polymorpic model, so there are a lot
of fields that are defined in the database that are not required
everywhere. To prevent having to add columns that aren’t relevant
everywhere, we add a new JSON column that will contain generic detail
about our Content Items. This also gives us flexibility to define new
Content Item types without having to further overload the database, or
create new models etc. It also gives us the flexibility to make creating
new content types a non-developer task in future if we want / need to.