Add config for the object store
In the next commit we’ll be adding an `ObjectStore::Item`
model, which will eventually have any number of fields.
Here we add some Yaml config to define the fields for
an `email_address` item type, and the fields that are
required for it.

We also add an `ObjectStore` configuration module, which
allows us to store / pull configuration in a meaningful
way, rather than pulling values out of hashes all over the
code.