Allow an item to be created or edited
This adds a new controller that inherits from the main
`EditionsController` and overrides some of the default behaviour
to allow our items to be edited/created.

We also add a new edit/new view to override the main editions
views which are already quite overloaded with logic.

At the moment, creation and editing fails when sending data to the
Publishing API, but this is a start.