Define list of delegated methods in Worldwide Organisations
We currently retrieve a list of methods that could be delegated to
contacts in the Worldwide Organisation model, then remove a number of
these.

As contacts haven't changed for many years, this dynamic code seems
unnecessary and is more difficult to understand than if the methods were
just hardcoded.

Therefore replacing with a list of methods. The delegation remains
tested in the `WorldwideOfficeTest`.

Also removes a comment about "quacking".