Fix translations not appearing on English draft when added
When a translation is added to an editionable document, we send the
draft of the translation to Publishing API. However we don't send the
draft of the English (or other translations), meaning the link to the
newly added translation doesn't appear.

This is an issue with all translatable editionable document types.

Therefore updating the controller to use `draft_updater` instead of
`draft_translation_updater`, which will send the draft of all
translations to Publishing API when a translation is updated.