Extract method for non editionable content types
and rename variable names to be a bit clearer what the two types of
republishable content types are,  as they're published differently based
on whether they're editionable or not.

We'll use the extracted method in the `BulkRepublisher` to determine
which republishing mechanism to use, as a different worker is used for
editionable and non-editionable document types.