Alter pre-election rake task to identify political content
Previously the task for marking political content for an organisation simply marked all of the content as political, whether it was eligible to be marked political or not.

The take task now applies the eligibility criteria when marking the content as political

This commit also fixes an ineffective test. Previously the test which checked that only documents published after the date provided was not associating the test document with the organisation effectively, meaning that the test would pass even if the date criteria was not applied.