Extract presenters to be called only when reshuffling
The logic for splitting functionality depending on reshuffling mode
used to live in the presenters themselves, checking `reshuffle_in_progress?`.

We have now moved that logic within `ReshuffleMode`, and are sending
content either to the draft or live stack depending on whether reshuffle
mode is on. We need the presenters to send full data (not empty objects) to
the required draft/live stack, so we are removing the checks from the presenters.

We have created new presenters to deal with sending blank links and body
for when we are enabling the reshuffle mode the first time.