Add preview link on the cabinet ministers ordering page
We did consider using the details component, which we use for
displating shareable preview links on editions etc, but that was
a bit misleading as it suggests a link that can be shared with
folks who don't have signon access. In this case, we're only
offering a shortcut to the draft-origin version of the ministers
index page, so it does require signon access, and the 'preview'
link here is not particularly shareable.

The code for the details component approach, should we choose to
use it in future, is available in
https://github.com/alphagov/whitehall/commit/36275690793e5b832b1f5300df993ee1db617628