Allow users to bulk republish by type
This is a fairly large commit, but it feels right to add all of this in
one go.

Here we add a page for selecting an editionable or non-editionable
content type for bulk republishing.  Selecting a content type takes you
to the familiar bulk republishing page where users enter a "reason" for
republishing.

Co-authored-by: Ynda <yndajas@gmail.com>