Add a base package
This can eventually contain all code that other packages depend on,
including base controllers, models and some components that are used
across the application. There are still some outstanding dependencies,
but we can refactor as we go.

We can now define the base package and dependencies in our packages and
run `packwerk update-todo`, which removes some entries from the todo
files.