Add GA4 tracking to all buttons on edition form
Introduce a new module to add ga4 data to all buttons contained within and include it in the edition form.
Clicks will then be tracked by the `ga4-event-tracker` which is already initialised.
For buttons that need custom event schemas, the module merges any existing attributes defined on the button with the generic set.