Make Consultation translateable
We've had a request to allow translations for the Consultation
format. Currently it is possible to create an English version
and a Foreign Language Only version of a Consultation, and
then manually link the two:
https://www.gov.uk/guidance/how-to-publish-on-gov-uk/translations#create-a-foreign-language-only-document-collection-or-consultation

However, we could equally support native translations in
Consultations, with a few lines of code as per this commit.

Admin view of translated Consultation:

![admin view of translated consultation](https://github.com/alphagov/whitehall/assets/5111927/922bb93f-42b4-444f-9494-726f4960da9b)

Consultation in English:

![test consultation English](https://github.com/alphagov/whitehall/assets/5111927/89ca6121-7f2c-4a91-acb5-30b81eaf1022)

Consultation translated into Azeri:

![test consultation Azeri](https://github.com/alphagov/whitehall/assets/5111927/46ef48b6-7b33-48b9-9b70-d9ea8eeef705)

But we have an issue - if we opt for a "foreign language only"
version of a Consultation, we can't add translations for it,
despite the UI suggesting we can.

Checking the foreign language only box:

![Checking the foreign language only box](https://github.com/alphagov/whitehall/assets/5111927/b2e957e1-21f8-48b9-8d75-5a74b57bc187)

It appears as a Welsh "translation":

![foreign language consultation admin view](https://github.com/alphagov/whitehall/assets/5111927/a55acdad-86d5-4540-9378-1cd9df385a78)

Notice that if we try to add a new translation, there's no English option:

![list of translation languages options - English is missing](https://github.com/alphagov/whitehall/assets/5111927/d7606402-4709-4de4-9065-d7b6406aa023)

If we do try to create a translation, we see a success message, but no translation is created (and no translation is published if you then force-publish the edition):

![success message "Arabic translation for 'Welsh consultation test' saved.". But no translation visible in the UI](https://github.com/alphagov/whitehall/assets/5111927/3cb3c86d-2110-41c8-a4f7-9a31ef46e840)

Zendesk ticket: https://govuk.zendesk.com/agent/tickets/5509163