Make Call for Evidence translateable
We've had a request to allow translations for the Call for
Evidence format. Currently it is possible to create an English
version, but there is no support for a "Foreign Language Only"
version of a Call for Evidence, so it is currently not possible
to publish a Call for Evidence in a different language.
This creates accessibility issues where publishers have to add
foreign language content but published as English.

Admin view of translated Call for Evidence:

![call for evidence UI with translations](https://github.com/alphagov/whitehall/assets/5111927/d23e3035-c017-4aba-a252-f82a3fb0a0ee)

Call for Evidence in English:

![call for evidence English](https://github.com/alphagov/whitehall/assets/5111927/b5d12731-3690-428e-b249-6b65c899620b)

Call for Evidence translated into Arabic:

![call for evidence Arabic](https://github.com/alphagov/whitehall/assets/5111927/e5d7a483-69e9-4815-a9d4-8ce42083429a)

Zendesk ticket: https://govuk.zendesk.com/agent/tickets/5509163