Add an Emergency Banner package
This adds the engine code in `lib` to define the engine, and moves the
code involved in the Emergency Banner (including the tests) to the
package directory, shaped as a mini-Rails app.

The routes for this package are defined in a routes file, and then
loaded in the main routes file with the `draw` method.

We also define a package in package.yml that tells Packwerk to enforce
dependencies. Running `packwerk update-todo` adds a `package_todo.yml`
file, which shows us where dependencies are being loaded in from other
parts of the application.