Add a base package
This can eventually contain all code that other packages depend on,
including base controllers, models and some components that are used
across the application.

We can now define the base package and dependencies in our package and
run `packwerk update-todo`, which removes the `package-todo` file in our
exiting package.

We also update our root application to enforce dependencies and
add the base package as a dependency. This ensures that the root
application doesn’t use any code in any of our packages, unless
a dependency is defined.