Fix exception handling in Rake task
With a deleted edition, this was raising a generic error, which is not
the intent expressed in the tests. The tests pass because the expected
error message was passed as a positional argument to `assert_raises`,
which is effectively ignored: it should instead be passed as a keyword
argument using `match`. This and similarly incorrect tests will be fixed
in a separate commit

Using the safe navigation operator, the intended exception-raising
behaviour of the rake task as documented in the tests is achieved for
deleted editions

---

This is needed because:
- the default scope of `Edition` excludes those with a state of
  `deleted` (from `Edition::Workflow`)
- `#attachable` is an association method inherited by `HtmlAttachment`
  from `Attachment`
- association methods appear to respect `default_scope`

So when the associated attachable is a deleted edition, it's out of
scope, leading `#attachable` to return `nil`. We still want to raise the
custom exception for deleted editions (per the tests), so we use the
safe navigation operator to get a falsey value (`nil`) when the edition
itself is `nil`