Fix error message matching in assert_raises
Some tests using `assert_raises` were including a string, presumably to
check that the error message looked correct. However, the
`assert_raises` method requires that this is a keyword argument using
`match`, which the tests were not using. This updates the tests to use
this, in some places fixing the expected error messages