add a basic Object Store engine
We want to create an engine to provide a way for interacting with re-usable content, that is namespaced away from the main Whitehall app in order to keep it contained, be 'good neighbours' and to keep in mind we may want to share its functionality with other publishing apps in the future.
This adds an empty module we are calling Object Store - we will shortly add models and controllers.