Add `content_ids` to `RepublishingEvent`
These will be used to capture the documents requested for republishing
for `all_documents_by_content_ids` events

One of the main reasons for capturing identifying information for bulk
republishing events that require extra input is to notify users when a
matching event was last triggered (this is a planned iteration of the
UI), so that they can make an informed decision on whether to request a
new republishing. This is perhaps less useful for this purpose, since we
could have partially matching arrays of content IDs across requests, and
we probably don't want to go as far as searching through all events of
this type to find the last appearance for each content ID in a list.
(For other republishing events that require extra input, there's a
single identifier for the whole event: a content ID, an organisation, or
a type.) However, this will still contribute to a detailed, searchable
record of which content has been requested for republishing and when

This is represented as JSON in the database as we're using MySQL2, which
doesn't support array columns