Make Publishing API calls directly instead of going via worker
We're seeing occasional race condition issues in Publishing API,
whereby the `payload_version` in Content Store is ahead of what's
being sent from Publishing API. We suspect this happens when
putting content, patching links and publishing content happens in
quick succession.

In hindsight, the patch/publish calls only need to happen when
reshuffle mode is toggled. The link patching only needs to happen
in the necessary callbacks, e.g. when ministers have been
reordered.

This commit reduces the number of calls we make to Publishing API,
and should mean we can delete a bunch of (now unused) code in the
next commit.