Allow users to bulk republish by content IDs
This adds a page where users can enter comma-, space-, and/or
new-line-separated content IDs, followed by a page where users are
presented with links to the editions that will republishing, asking
users to provide a reason and confirm that they want the editions to be
republished. Users are then redirected back to the index with a flashed
confirmation message

In the `confirm_documents_by_content_ids` view, we use a custom title
and page title. The reason for this is that we can't make a reliable
assumption about how many content IDs users will try republishing at
once, and if there are more than one or two, the titles will get too
long. We still want a unique page title, so to avoid using the same name
as the `new_document_by_content_ids` view, here we use "all documents by
given content IDs"/"all documents with the given content IDs" instead:
unique with a fixed length

As a result of this, we don't need to send the bulk content type
metadata down from the controller action to the view, so it's not stored
in an instance variable. However, the name `bulk_content_type_metadata`
is already taken by the method, so convention is broken and this is
called `bulk_content_type_metadata_for_content_ids` by exception