Remove unused `save_draft` method
This was a nice idea in principle, but the only use case we had
for it was for creating a preview version of the ministers page
after publishing a maintenance message on the live stack. The
nature of using workers for this meant that they could run out of
order and we'd end up overriding the maintenance message on live.

We've now changed how we apply the live and draft payloads, and
this method is no longer called. Rather than keep it 'just in
case' something in the future wants it, we'll delete it and this
commit can be reverted in future if a need for it arises.