Patch links rather than republish ministers index page
Republish isn't needed, as the only content that would be getting
published is the bit of body text at the top of the page, which
won't have changed. What's changed (in all of the callbacks that
trigger this) are the links describing the ministers.
`patch_links` is the way forward here, to update the content item:
the page will reflect the changes because it renders the links
dynamically.