Remove empty ministerial roles from drag-n-drop
In preparation for the election, we are improving the ordering of
cabinet ministers. The empty, unappointed roles, were showing in the
ordering page (and during drag and drop), which made the process
quite cumbersome.

We filtered out the empty roles to streamline the reordering task.
The roles will only show if they have active, current appointments.
The roles still show in the roles page.
Other tabs on the ordering page are not affected (e.g. 'Also
attends cabinet ordering').

Co-authored-by: Kashif Atcha <kashif.atcha@digital.cabinet-office.gov.uk>
Co-authored-by: Ryan Brown <ryan.brown@digital.cabinet-office.gov.uk>