Prevent accidental republication of unsafe documents
The republish all documents of a given type rake task failed to apply the guard function checking that documents are safe to republish.

This resulted in old statistics announcements being inadvertently republished, replacing the redirect from the announcements to their published statistics publications.

This commit prevents accidental republication of models that include the PublishesToPublishingApi module