Add GA4 tracking on page load
Sending data related to the current user and page.
We are sending a page_view event, with custom fields.

In GTM, it seems that the govuk_publishing_components page_view event
fields merge with the custom ones, and only one page_view event is
produced on page_load, which is exactly what we wanted.

Co-authored-by: Kashif Atcha <kashif.atcha@digital.cabinet-office.gov.uk>