Add a basic Object Store engine
We want to create an engine to provide a way for interacting with re-usable content, that is namespaced away from the main Whitehall app in order to keep it contained, be 'good neighbours' and to keep in mind we may want to share its functionality with other publishing apps in the future.

This adds an empty module we are calling “Content Object Store”. The wider Content Modelling and Reuse team are currently referring to this as the “object store”, however we think in the context of Whitehall this naming is too generic and so prefix it with “content” to help readers of the code.

Co-authored-by: Harriet H-W <harriethw@gmail.com>
Co-authored-by: pezholio <pezholio@gmail.com>