Create a minimal Content Block Edition
We take the minimal fields needed to imitate an Edition.

In this case we have added
a `details` json column which we assume will contain the properties of our block,
for example `{ "email_address": "some@example.com" }`. We are using `details`
here as that's the key used in the Publishing API to publish a piece of content[1]

We are taking the approach of duplication before reuse to help us move quickly and learn where the parts we would like to reuse are.

[1] https://govuk-pact-broker-6991351eca05.herokuapp.com/pacts/provider/Publishing%20API/consumer/GDS%20API%20Adapters/latest#a_request_from_the_Whitehall_application_to_create_a_content_item_at_/test-item_given_/test-item_has_been_reserved_by_the_Publisher_application

Co-authored-by: Tom Hipkin <tomhipkin@gmail.com>