Allow Content Blocks to be created
This adds a simple form that uses all a schema’s fields and saves all
the user input. This is very basic for now, but allows us to prove the
concept.

There’s also no validation, but that can come later

I’ve tried to introduce the concept of Cucumber-like steps in the
integration tests as well.