Republish statistical publications for selected political organisations
A number of statistical publications have been identified on GOV.UK that have history mode enabled. Statistical publications are not eligible for history mode, but a bug in a previous version of the election:identify_political_content_for rake task marked them as political incorrectly. Some of the incorrectly marked editions were published to GOV.UK.

The offending statistical publications have since had the political flag removed, but have not yet been republished. Republishing the documents should result in the removal of the history mode banner from the documents.