Introduce a new CreateContentEdition service
This will take care of the two responsibilities of creating an edition in two different places whilst removing this logic from the controller level.

We wrap the two calls in a database transaction as if the local request succeeds but the remote fails, we want to go back to a clean starting point.

We stub the schemas allowed in the test context to expect ‘block_type’. It doesn’t matter what this value is as long as it doesn’t couple us to a a real production type that may be removed.