Add an access layer when getting the valid schemas for stubbing
We want to test with a fake block type “block_type” instead of coupling ourselves to a real production type eg. “email_address”.

With the previous frozen constant we could not add a new type into the testing context.

What we can do is add an accessor method which we can stub so we don’t mess with the contents of the constant.

We finally add a unit test that confirms the block types in the constant and the method are those we return. This duplicates the list but hopefully it’s straight forward enough.