Tidy up delegation
When we were creating an edition, we were using `attr_accessor` values
to persist a Document’s title (using `document_title`) and block type.

This commit tidies up how we delegate to documents, using `attr_writer`
to write to instance variables when creating an edition, and then using
custom getters for the title and block type. If a document exists and
has the attribute, then we return the attribute from the document. If it
doesn’t, we attempt to return the value of the instance variable.

I’ve also changed the slightly clumsy wording of `document_title` to
just `title`, so we’re more consistent.