Change `worldwide_organisations` relationship to use new model
We have replaced `WorldwideOrganisation` with
`EditionableWorldwideOrganisation`, so need to return the editionable
version when querying which roles have worldwide organisations.

Note we have swapped the assignment during the migration. Previously
Worldwide Organisations were assigned to Roles. Now Roles are assigned
to Worldwide Organisations. This is why the relationship name (and link
table) are named differently.