Refactor the show content block helpers so a changed title can be injected
For new content block edition objects it worked to pass the object along and then ask it for the document title in order to make assertions.

However now we want to use the same helper for changing the document title to a new value. We now want to compare the object against a new value that we declare in the tests. We will see this used in the following commits.