ContentBlockEdition only validates a block_type on create
When we __create__ a ContentBlockEdition we need the block_type and details to create both the Edition and the Document.

When we want to __update__ the Editions details or the Documents titles we want to do so without changing the block_type. Changing the block type would change the schema of the edition and that would break the integrity of the data within.

We could pass the block_type along from the view as a hidden field or manually add it in the controller. The first can be tampered with and the second is arguably confusing, why go out of the way to provide a value that always has the be the previous value?