Bug fix - Only delete removed translations upon publish or force publish
At the moment, on every action of the publishing API pusher, deleted translations are unpublished from Publishing API. This has caused some bugs where live translations were deleted when only the draft has been updated. This is due to the unpublish-gone redirect taking only a content ID, and hence unable to differentiate between removing the draft or removing the live edition on Publishing API. The draft of the translation that has been deleted is removed from Publishing API already during the translation deletion step.

This change will fix live translations being pulled prematurely. The only problem we have left to solve is the preview not reflecting exactly the correct translation status of the new draft edition.